Implemented read-only properties.
Removing the set property method for read-only properties requiring that you set those properties in the constructor. Another way to do it would be to keep track if the entity has been saved or not, and throw an exception when you try to set the read-only property of a persisted entity, but I like to have the check in compile time instead of runtime.

Read-only properties are optional in the constructor, you don't have to set them, unless they are also mandatory. And yet, if they are read-only, mandatory but have a default value, they are still optional.